% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibble.R
\name{as_funtsibble}
\alias{as_funtsibble}
\alias{as_funtsibble.demogdata}
\alias{as_funtsibble.fts}
\title{Coerce to a funtsibble object}
\usage{
as_funtsibble(x, key = NULL, index, regular = TRUE, validate = TRUE,
  .drop = TRUE, ...)

\method{as_funtsibble}{demogdata}(x, ..., validate = TRUE)

\method{as_funtsibble}{fts}(x, ..., validate = TRUE)
}
\arguments{
\item{x}{Objects to be coerced to a funtsibble (`fun_tbl_ts`).}

\item{validate}{`TRUE` suggests to verify that each key or each combination
of key variables leads to unique time indices (i.e. a valid tsibble). If you
are sure that it's a valid input, specify `FALSE` to skip the checks.}

\item{...}{Other arguments passed on to individual methods.}
}
\value{
A funtsibble object.
}
\description{
Coerce to a funtsibble object
}
\examples{
# coerce demogdata object to tsibble ----
as_funtsibble(demography::fr.mort)
# coerce demogdata object to tsibble ----
as_funtsibble(rainbow::ElNino_ERSST_region_1and2)
}
\seealso{
\code{\link[tsibble]{as_tsibble}}
}
